refactor config to separate file
